refactor to avoid Kryo exceptions
